+++ continuous_integration/install.sh
@@ -76,7 +76,7 @@ python -m pip install -q -r dev-requirements.txt
 bash ./continuous_integration/build_test_ext.sh
 
 python --version
-python -c "import numpy; print('numpy %s' % numpy.__version__)" || echo "no numpy"
-python -c "import scipy; print('scipy %s' % scipy.__version__)" || echo "no scipy"
+python -c "import numpy; print(f'numpy {numpy.__version__}')" || echo "no numpy"
+python -c "import scipy; print(f'scipy {scipy.__version__}')" || echo "no scipy"
 
 python -m flit install --symlink