+++ tests/test_threadpoolctl.py
@@ -5,7 +5,8 @@ import re
 import subprocess
 import sys
 
-from threadpoolctl import threadpool_limits, threadpool_info, _ThreadpoolInfo
+from threadpoolctl import threadpool_limits, threadpool_info
+from threadpoolctl import ThreadpoolController
 from threadpoolctl import _ALL_PREFIXES, _ALL_USER_APIS
 
 from .utils import cython_extensions_compiled
@@ -14,10 +15,10 @@ from .utils import scipy
 from .utils import threadpool_info_from_subprocess
 
 
-def is_old_openblas(module):
+def is_old_openblas(lib_controller):
     # Possible bug in getting maximum number of threads with OpenBLAS < 0.2.16
     # and OpenBLAS does not expose its version before 0.3.4.
-    return module.internal_api == "openblas" and module.version is None
+    return lib_controller.internal_api == "openblas" and lib_controller.version is None
 
 
 def effective_num_threads(nthreads, max_threads):
@@ -26,130 +27,191 @@ def effective_num_threads(nthreads, max_threads):
     return nthreads
 
 
-def _threadpool_info():
-    # Like threadpool_info but return the object instead of the list of dicts
-    return _ThreadpoolInfo(user_api=_ALL_USER_APIS)
+def test_threadpool_info():
+    # Check consistency between threadpool_info and ThreadpoolController
+    function_info = threadpool_info()
+    object_info = ThreadpoolController().lib_controllers
+
+    for lib_info, lib_controller in zip(function_info, object_info):
+        assert lib_info == lib_controller.info()
 
 
-def test_threadpool_limits_public_api():
-    # Check consistency between threadpool_info and _ThreadpoolInfo
-    public_info = threadpool_info()
-    private_info = _threadpool_info()
+def test_threadpool_controller_info():
+    # Check that all keys expected for the private api are in the dicts
+    # returned by the `info` methods
+    controller = ThreadpoolController()
 
-    for module1, module2 in zip(public_info, private_info):
-        assert module1 == module2.todict()
+    assert threadpool_info() == [
+        lib_controller.info() for lib_controller in controller.lib_controllers
+    ]
+    assert controller.info() == [
+        lib_controller.info() for lib_controller in controller.lib_controllers
+    ]
+
+    for lib_controller_dict in controller.info():
+        assert "user_api" in lib_controller_dict
+        assert "internal_api" in lib_controller_dict
+        assert "prefix" in lib_controller_dict
+        assert "filepath" in lib_controller_dict
+        assert "version" in lib_controller_dict
+        assert "num_threads" in lib_controller_dict
+
+        if lib_controller_dict["internal_api"] in ("mkl", "blis", "openblas"):
+            assert "threading_layer" in lib_controller_dict
 
 
-def test_ThreadpoolInfo_todicts():
-    # Check all keys expected for the public api are in the dicts returned by
-    # the .todict(s) methods
-    info = _threadpool_info()
+@pytest.mark.parametrize(
+    "kwargs",
+    [
+        {"user_api": "blas"},
+        {"prefix": "libgomp"},
+        {"internal_api": "openblas", "prefix": "libomp"},
+        {"prefix": ["libgomp", "libomp", "libiomp"]},
+    ],
+)
+def test_threadpool_controller_select(kwargs):
+    # Check the behior of the select method of ThreadpoolController
+    controller = ThreadpoolController().select(**kwargs)
+    if not controller:
+        pytest.skip(f"Requires at least one of {list(kwargs.values())}.")
 
-    assert threadpool_info() == [module.todict() for module in info.modules]
-    assert info.todicts() == [module.todict() for module in info]
-    assert info.todicts() == [module.todict() for module in info.modules]
-
-    for module in info:
-        module_dict = module.todict()
-        assert "user_api" in module_dict
-        assert "internal_api" in module_dict
-        assert "prefix" in module_dict
-        assert "filepath" in module_dict
-        assert "version" in module_dict
-        assert "num_threads" in module_dict
-
-        if module.internal_api in ("mkl", "blis", "openblas"):
-            assert "threading_layer" in module_dict
+    for lib_controller in controller.lib_controllers:
+        assert any(
+            getattr(lib_controller, key) in (val if isinstance(val, list) else [val])
+            for key, val in kwargs.items()
+        )
 
 
 @pytest.mark.parametrize("prefix", _ALL_PREFIXES)
 @pytest.mark.parametrize("limit", [1, 3])
 def test_threadpool_limits_by_prefix(prefix, limit):
     # Check that the maximum number of threads can be set by prefix
-    original_info = _threadpool_info()
+    original_controller = ThreadpoolController()
 
-    modules_matching_prefix = original_info.get_modules("prefix", prefix)
-    if not modules_matching_prefix:
+    controller_matching_prefix = original_controller.select(prefix=prefix)
+    if not controller_matching_prefix:
         pytest.skip(f"Requires {prefix} runtime")
 
     with threadpool_limits(limits={prefix: limit}):
-        for module in modules_matching_prefix:
-            if is_old_openblas(module):
+        for lib_controller in controller_matching_prefix.lib_controllers:
+            if is_old_openblas(lib_controller):
                 continue
             # threadpool_limits only sets an upper bound on the number of
             # threads.
-            assert 0 < module.get_num_threads() <= limit
-    assert _threadpool_info() == original_info
+            assert 0 < lib_controller.get_num_threads() <= limit
+    assert ThreadpoolController().info() == original_controller.info()
 
 
 @pytest.mark.parametrize("user_api", (None, "blas", "openmp"))
 @pytest.mark.parametrize("limit", [1, 3])
 def test_set_threadpool_limits_by_api(user_api, limit):
     # Check that the maximum number of threads can be set by user_api
-    original_info = _threadpool_info()
+    original_controller = ThreadpoolController()
 
-    modules_matching_api = original_info.get_modules("user_api", user_api)
-    if not modules_matching_api:
+    if user_api is None:
+        controller_matching_api = original_controller
+    else:
+        controller_matching_api = original_controller.select(user_api=user_api)
+    if not controller_matching_api:
         user_apis = _ALL_USER_APIS if user_api is None else [user_api]
         pytest.skip(f"Requires a library which api is in {user_apis}")
 
     with threadpool_limits(limits=limit, user_api=user_api):
-        for module in modules_matching_api:
-            if is_old_openblas(module):
+        for lib_controller in controller_matching_api.lib_controllers:
+            if is_old_openblas(lib_controller):
                 continue
             # threadpool_limits only sets an upper bound on the number of
             # threads.
-            assert 0 < module.get_num_threads() <= limit
+            assert 0 < lib_controller.get_num_threads() <= limit
 
-    assert _threadpool_info() == original_info
+    assert ThreadpoolController().info() == original_controller.info()
 
 
 def test_threadpool_limits_function_with_side_effect():
     # Check that threadpool_limits can be used as a function with
     # side effects instead of a context manager.
-    original_info = _threadpool_info()
+    original_controller = ThreadpoolController()
 
     threadpool_limits(limits=1)
     try:
-        for module in _threadpool_info():
-            if is_old_openblas(module):
+        for lib_controller in ThreadpoolController().lib_controllers:
+            if is_old_openblas(lib_controller):
                 continue
-            assert module.num_threads == 1
+            assert lib_controller.num_threads == 1
     finally:
         # Restore the original limits so that this test does not have any
         # side-effect.
-        threadpool_limits(limits=original_info)
+        threadpool_limits(limits=original_controller)
 
-    assert _threadpool_info() == original_info
+    assert ThreadpoolController().info() == original_controller.info()
 
 
 def test_set_threadpool_limits_no_limit():
     # Check that limits=None does nothing.
-    original_info = _threadpool_info()
+    original_controller = ThreadpoolController()
     with threadpool_limits(limits=None):
-        assert _threadpool_info() == original_info
+        assert ThreadpoolController().info() == original_controller.info()
 
-    assert _threadpool_info() == original_info
+    assert ThreadpoolController().info() == original_controller.info()
 
 
 def test_threadpool_limits_manual_unregister():
     # Check that threadpool_limits can be used as an object which holds the
     # original state of the threadpools and that can be restored thanks to the
     # dedicated unregister method
-    original_info = _threadpool_info()
+    original_controller = ThreadpoolController()
 
     limits = threadpool_limits(limits=1)
     try:
-        for module in _threadpool_info():
-            if is_old_openblas(module):
+        for lib_controller in ThreadpoolController().lib_controllers:
+            if is_old_openblas(lib_controller):
                 continue
-            assert module.num_threads == 1
+            assert lib_controller.num_threads == 1
     finally:
         # Restore the original limits so that this test does not have any
         # side-effect.
         limits.unregister()
 
-    assert _threadpool_info() == original_info
+    assert ThreadpoolController().info() == original_controller.info()
+
+
+def test_threadpool_controller_limit():
+    # Check that using the limit method of ThreadpoolController only impact its
+    # library controllers.
+    original_blas_controller = ThreadpoolController().select(user_api="blas")
+    original_openmp_controller = ThreadpoolController().select(user_api="openmp")
+
+    with original_blas_controller.limit(limits=1):
+        blas_controller = ThreadpoolController().select(user_api="blas")
+        openmp_controller = ThreadpoolController().select(user_api="openmp")
+
+        assert all(
+            lib_controller.num_threads == 1
+            for lib_controller in blas_controller.lib_controllers
+        )
+        # original_blas_controller contains only blas libraries so no opemp library
+        # should be impacted.
+        assert openmp_controller.info() == original_openmp_controller.info()
+
+
+def test_threadpool_controller_restore():
+    # Check that the restore_limits method of ThreadpoolController is able to set the
+    # limits back to their original values. Similar to
+    # test_threadpool_limits_function_with_side_effect but with the object api
+    controller = ThreadpoolController()
+
+    controller.limit(limits=1)
+    try:
+        for lib_controller in ThreadpoolController().lib_controllers:
+            if is_old_openblas(lib_controller):
+                continue
+            assert lib_controller.num_threads == 1
+    finally:
+        # Restore the original limits so that this test does not have any
+        # side-effect.
+        controller.restore_limits()
+
+    assert ThreadpoolController().info() == controller.info()
 
 
 def test_threadpool_limits_bad_input():
@@ -211,11 +273,11 @@ def test_openmp_nesting(nthreads_outer):
         # There are 2 openmp, the one from inner and the one from outer.
         assert len(outer_info) == 2
         # We already know the one from inner. It has to be the other one.
-        prefixes = {module["prefix"] for module in outer_info}
+        prefixes = {lib_info["prefix"] for lib_info in outer_info}
         outer_omp = prefixes - {inner_omp}
 
     outer_num_threads, inner_num_threads = check_nested_openmp_loops(10)
-    original_info = _threadpool_info()
+    original_controller = ThreadpoolController()
 
     if inner_omp == outer_omp:
         # The OpenMP runtime should be shared by default, meaning that the
@@ -232,7 +294,7 @@ def test_openmp_nesting(nthreads_outer):
 
     # The state of the original state of all threadpools should have been
     # restored.
-    assert _threadpool_info() == original_info
+    assert ThreadpoolController().info() == original_controller.info()
 
     # The number of threads available in the outer loop should not have been
     # decreased:
@@ -245,7 +307,7 @@ def test_openmp_nesting(nthreads_outer):
             # XXX: this does not always work when nesting independent openmp
             # implementations. See: https://github.com/jeremiedbb/Nested_OpenMP
             pytest.xfail(
-                "Inner OpenMP num threads was %d instead of 1" % inner_num_threads
+                f"Inner OpenMP num threads was {inner_num_threads} instead of 1"
             )
     assert inner_num_threads == 1
 
@@ -253,15 +315,15 @@ def test_openmp_nesting(nthreads_outer):
 def test_shipped_openblas():
     # checks that OpenBLAS effectively uses the number of threads requested by
     # the context manager
-    original_info = _threadpool_info()
+    original_controller = ThreadpoolController()
 
-    openblas_modules = original_info.get_modules("internal_api", "openblas")
+    openblas_controllers = original_controller.select(internal_api="openblas")
 
     with threadpool_limits(1):
-        for module in openblas_modules:
-            assert module.get_num_threads() == 1
+        for lib_controller in openblas_controllers.lib_controllers:
+            assert lib_controller.get_num_threads() == 1
 
-    assert original_info == _threadpool_info()
+    assert original_controller.info() == ThreadpoolController().info()
 
 
 @pytest.mark.skipif(
@@ -288,17 +350,20 @@ def test_nested_prange_blas(nthreads_outer):
 
     check_nested_prange_blas = prange_blas.check_nested_prange_blas
 
-    original_info = _threadpool_info()
+    original_controller = ThreadpoolController()
 
-    blas_info = original_info.get_modules("user_api", "blas")
-    blis_info = original_info.get_modules("internal_api", "blis")
+    blas_controllers = original_controller.select(user_api="blas")
+    blis_controllers = original_controller.select(internal_api="blis")
 
     # skip if the BLAS used by numpy is an old openblas. OpenBLAS 0.3.3 and
     # older are known to cause an unrecoverable deadlock at process shutdown
     # time (after pytest has exited).
     # numpy can be linked to BLIS for CBLAS and OpenBLAS for LAPACK. In that
     # case this test will run BLIS gemm so no need to skip.
-    if not blis_info and any(is_old_openblas(module) for module in blas_info):
+    if not blis_controllers and any(
+        is_old_openblas(lib_controller)
+        for lib_controller in blas_controllers.lib_controllers
+    ):
         pytest.skip("Old OpenBLAS: skipping test to avoid deadlock")
 
     A = np.ones((1000, 10))
@@ -309,17 +374,19 @@ def test_nested_prange_blas(nthreads_outer):
         nthreads = effective_num_threads(nthreads_outer, max_threads)
 
         result = check_nested_prange_blas(A, B, nthreads)
-        C, prange_num_threads, inner_info = result
+        C, prange_num_threads, inner_controller = result
 
     assert np.allclose(C, np.dot(A, B.T))
     assert prange_num_threads == nthreads
 
-    nested_blas_info = inner_info.get_modules("user_api", "blas")
-    assert len(nested_blas_info) == len(blas_info)
-    for module in nested_blas_info:
-        assert module.num_threads == 1
+    nested_blas_controllers = inner_controller.select(user_api="blas")
+    assert len(nested_blas_controllers.lib_controllers) == len(
+        blas_controllers.lib_controllers
+    )
+    for lib_controller in nested_blas_controllers.lib_controllers:
+        assert lib_controller.num_threads == 1
 
-    assert original_info == _threadpool_info()
+    assert original_controller.info() == ThreadpoolController().info()
 
 
 # the method `get_original_num_threads` raises a UserWarning due to different
@@ -330,20 +397,23 @@ def test_nested_prange_blas(nthreads_outer):
 @pytest.mark.parametrize("limit", [1, None])
 def test_get_original_num_threads(limit):
     # Tests the method get_original_num_threads of the context manager
-    with threadpool_limits(limits=2, user_api="blas") as ctl:
+    with threadpool_limits(limits=2, user_api="blas") as ctx:
         # set different blas num threads to start with (when multiple openblas)
-        if ctl._original_info:
-            ctl._original_info.modules[0].set_num_threads(1)
+        if ctx._controller:
+            ctx._controller.lib_controllers[0].set_num_threads(1)
 
-        original_info = _threadpool_info()
+        original_controller = ThreadpoolController()
         with threadpool_limits(limits=limit, user_api="blas") as threadpoolctx:
             original_num_threads = threadpoolctx.get_original_num_threads()
 
             assert "openmp" not in original_num_threads
 
-            blas_info = original_info.get_modules("user_api", "blas")
-            if blas_info:
-                expected = min(module.num_threads for module in blas_info)
+            blas_controller = original_controller.select(user_api="blas")
+            if blas_controller:
+                expected = min(
+                    lib_controller.num_threads
+                    for lib_controller in blas_controller.lib_controllers
+                )
                 assert original_num_threads["blas"] == expected
             else:
                 assert original_num_threads["blas"] is None
@@ -356,30 +426,30 @@ def test_get_original_num_threads(limit):
 def test_mkl_threading_layer():
     # Check that threadpool_info correctly recovers the threading layer used
     # by mkl
-    mkl_info = _threadpool_info().get_modules("internal_api", "mkl")
+    mkl_controller = ThreadpoolController().select(internal_api="mkl")
     expected_layer = os.getenv("MKL_THREADING_LAYER")
 
-    if not (mkl_info and expected_layer):
+    if not (mkl_controller and expected_layer):
         pytest.skip("requires MKL and the environment variable MKL_THREADING_LAYER set")
 
-    actual_layer = mkl_info.modules[0].threading_layer
+    actual_layer = mkl_controller.lib_controllers[0].threading_layer
     assert actual_layer == expected_layer.lower()
 
 
 def test_blis_threading_layer():
     # Check that threadpool_info correctly recovers the threading layer used
     # by blis
-    blis_info = _threadpool_info().get_modules("internal_api", "blis")
+    blis_controller = ThreadpoolController().select(internal_api="blis")
     expected_layer = os.getenv("BLIS_ENABLE_THREADING")
     if expected_layer == "no":
         expected_layer = "disabled"
 
-    if not (blis_info and expected_layer):
+    if not (blis_controller and expected_layer):
         pytest.skip(
             "requires BLIS and the environment variable BLIS_ENABLE_THREADING set"
         )
 
-    actual_layer = blis_info.modules[0].threading_layer
+    actual_layer = blis_controller.lib_controllers[0].threading_layer
     assert actual_layer == expected_layer
 
 
@@ -395,8 +465,8 @@ def test_libomp_libiomp_warning(recwarn):
 
     # Check that a warning is raised when both libomp and libiomp are loaded
     # It should happen in one CI job (pylatest_conda_mkl_clang_gcc).
-    info = _threadpool_info()
-    prefixes = [module.prefix for module in info]
+    controller = ThreadpoolController()
+    prefixes = [lib_controller.prefix for lib_controller in controller.lib_controllers]
 
     if not ("libomp" in prefixes and "libiomp" in prefixes and sys.platform == "linux"):
         pytest.skip("Requires both libomp and libiomp loaded, on Linux")
@@ -422,8 +492,8 @@ def test_command_line_command_flag():
     cli_info = json.loads(output.decode("utf-8"))
 
     this_process_info = threadpool_info()
-    for module in cli_info:
-        assert module in this_process_info
+    for lib_info in cli_info:
+        assert lib_info in this_process_info
 
 
 @pytest.mark.skipif(
@@ -448,8 +518,8 @@ def test_command_line_import_flag():
     cli_info = json.loads(result.stdout)
 
     this_process_info = threadpool_info()
-    for module in cli_info:
-        assert module in this_process_info
+    for lib_info in cli_info:
+        assert lib_info in this_process_info
 
     warnings = [w.strip() for w in result.stderr.splitlines()]
     assert "WARNING: could not import invalid_package" in warnings
@@ -473,11 +543,11 @@ def test_architecture():
         "skx",
         "haswell",
     )
-    for module in threadpool_info():
-        if module["internal_api"] == "openblas":
-            assert module["architecture"] in expected_openblas_architectures
-        elif module["internal_api"] == "blis":
-            assert module["architecture"] in expected_blis_architectures
+    for lib_info in threadpool_info():
+        if lib_info["internal_api"] == "openblas":
+            assert lib_info["architecture"] in expected_openblas_architectures
+        elif lib_info["internal_api"] == "blis":
+            assert lib_info["architecture"] in expected_blis_architectures
         else:
-            # Not supported for other modules
-            assert "architecture" not in module
+            # Not supported for other libraries
+            assert "architecture" not in lib_info